backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
python-multipart
chromadb
sentence-transformers
alembic
google-genai